Add Means to Get Enterprise Year Data and Display them on Screen
Get the selected entreprise (from the combobox) corresponding year data from the year table and display them on identification tab and by selecting the year from the combo box.

Added a stored procedure to select entreprise  year data by selected enterprise id, added model and data access class for the Entr year in the db Library, a controller in the data manager a api helper in the ui library to connect to the endpoint, now we can load the entreprise year data in the view, by selecting the year item from the combobox.